requirements.txt
flask==2.3.3
flask-session==0.5.0
pyodbc==4.0.39
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==2.3.7